refactor: share auth response contracts
