+++ api/_utils/auth/_types.ts
@@ -45,24 +45,9 @@ export interface AuthenticatedUser {
 // Response Types
 // ============================================================================
 
-export interface AuthErrorResponse {
-  error: string;
-  message?: string;
-}
-
-export interface TokenResponse {
-  token: string;
-  username?: string;
-}
-
-export interface VerifyTokenResponse {
-  valid: boolean;
-  username: string;
-  message: string;
-  expired?: boolean;
-}
-
-export interface CheckPasswordResponse {
-  hasPassword: boolean;
-  username: string;
-}
+export type {
+  AuthErrorResponse,
+  TokenResponse,
+  VerifyTokenResponse,
+  CheckPasswordResponse,
+} from "../../../src/shared/contracts/auth.js";
